refactor: enforce widget file line limits
