+++ src/widget/capture-timeout.ts
@@ -1,0 +1,13 @@
+const CAPTURE_TIMEOUT_MS = 15_000;
+
+export function withCaptureTimeout<T>(capturePromise: Promise<T>): Promise<T> {
+  let timer: ReturnType<typeof setTimeout>;
+  const timeoutPromise = new Promise<never>((_, reject) => {
+    timer = setTimeout(
+      () => reject(new Error('Screenshot capture timed out — the page may be too complex')),
+      CAPTURE_TIMEOUT_MS
+    );
+  });
+
+  return Promise.race([capturePromise, timeoutPromise]).finally(() => clearTimeout(timer!));
+}
